Finished task 3
I separated out the testing of the bank contents into a new thread,
had that thread start at the saem time as all other threads, and then
made sure that the new testing thread would stop when the bank is
"closed". The bank is "closed" when any one thread is finished looping
through all of it's transactions.